# swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Diploma Project
  version: 1.0.0
...
paths:
...
  /catalog/:
    get:
      tags:
        - catalog
      description: 'get catalog items'
      parameters:
        - name: filter
          in: query
          description: search text
          required: false
          schema:
            type: string
            default: ''
        - name: page
          in: query
          description: page
          required: false
          schema:
            type: string
            default: ''
        - name: category
          in: query
          required: false
          schema:
            type: string
            default: ''
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: date
            enum:
              - rating
              - price
              - reviews
              - date
        - name: sortType
          in: query
          required: false
          schema:
            type: string
            default: dec
            enum:
              - dec
              - inc
        - name: limit
          in: query
          required: false
          schema:
            type: number
            default: 20

      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductShort'
                  currentPage:
                    type: number
                    example: 5
                  lastPage:
                    type: number
                    example: 10
  /catalog/{id}/:
    get:
      tags:
        - catalog
      description: 'get catalog items'
      parameters:
        - name: filter
          in: query
          description: search text
          required: false
          schema:
            type: string
            default: ''
        - name: page
          in: query
          description: page
          required: false
          schema:
            type: string
            default: ''
        - name: category
          in: query
          required: false
          schema:
            type: string
            default: ''
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: date
            enum:
              - rating
              - price
              - reviews
              - date
        - name: sortType
          in: query
          required: false
          schema:
            type: string
            default: dec
            enum:
              - dec
              - inc
        - name: limit
          in: query
          required: false
          schema:
            type: number
            default: 20
        - name: id
          in: path
          description: category id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductShort'
                  currentPage:
                    type: number
                    example: 5
                  lastPage:
                    type: number
                    example: 10
...
components:
  schemas:
...
    ProductShort:
      required:
        - id
        - categoryName
        - title
        - price
        - href
      type: object
      xml:
        name: productShort
      properties:
        id:
          type: integer
          example: 123
        categoryName:
          type: string
          example: 'Video cards'
        price:
          type: number
          example: 500.67
        count:
          type: integer
          example: 12
          nullable: true
        date:
          type: string
          example: 'Thu Feb 09 2023 21:39:52 GMT+0100 (Central European Standard Time)'
          nullable: true
        title:
          type: string
          example: 'NVIDIA GeForce RTX'
        description:
          type: string
          example: 'description of the product'
        href:
          type: string
          example: '/product/123'
        freeDelivery:
          type: boolean
          example: true
          nullable: true
        images:
          type: array
          xml:
            wrapped: true
          items:
            type: string
            example: '/media/products/product_1/images/example.jpg'
            xml:
              name: imageUrl
        tags:
          type: array
          xml:
            wrapped: true
          items:
            type: string
            xml:
              name: tagId
        reviews:
          type: integer
          example: 5
          nullable: true
        rating:
          type: number
          example: 4.6
...